feat: function to produce a signed token
Includes a small amount of decoding for verification, that test can
probably be rewritten to use final decoding code in the next ticket.

Also includes test keys sourced out of other public edx tests.  The
"wrong key" may or may not actually be valid, it is included just to
validate that the correct key of multiple is picked up during decode
(because the key ID ends up in the token)